+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.0.0
+
+- Fixes eslint plugin not reporting errors on `PropTypes.any` and `PropTypes.symbol`
+- Fixes eslint plugin not reporting errors on identifiers in meta
+
 ## 0.2.0
 
 - BREAKING: [Compiler, Webpack plugin] Renames `baseClass` compiler option to `supertype`